Add new option `max_y_floor`
Allow users to set a minimum for the y-axis max value. In contrast to
`max_y`, `max_y_floor` allows the y-axis to rescale when new data is
supplied instead of locking the y-axis max to a fixed value.
`max_y_floor` is the smallest value that the y-axis max can take.