Handle years-long time differences better when drawing x axis (fixes #848)
Increase the threshold at which we display the x axis in terms of years
(for ranges of around 2 years, this could mean that we display the same
year multiple times on the x axis, due to the way d3 decides on the
number of ticks).

This required some follow-on changes to the way we decide to add a
secondary year marker to the x axis.