Make it harder to pass in data to metricsgraphics line charts that gets clobbered (fixes #896)
We set index and line_id properties on the data passed in so we can do some
lookups on these objects later. This isn't optimal, but is hard to fix
properly given the current state of the code. Workaround this by wrapping
the properties with `__`, making it less likely they'll get clobbered.